refactor(canvas): extract chat attachment domain
