+++ apps/canvas-workspace/src/renderer/src/agent-chat/attachments/attachmentFileName.ts
